requirements.txt
feedparser>=6.0
google-generativeai>=0.7
pyyaml>=6.0
jinja2>=3.1
slack-sdk>=3.0
python-dotenv>=1.0
